Add work-in-progress code to store the particles and parton jets
This works, but isn't great. I think I will rewrite this to be more like
the other output pieces of the analysis framework-- probably the right
thing to do is to have a single class which stores particles and parton
jets for a given status code, and just create as many of these are
needed.