Work towards implementing Sherpa parton clustering algorithm.
The skeleton of the algorithm is now in place, though I haven't
really tested it yet. There are currently two helper classes
to the main algorithm:

* PartonClusterer: given some truth particles, writes them
to the tree, then clusters them and produces parton jets.
One of these gets created for status 20 and status 3 particles,
and they contain the other helper class:

* DijetFinder: given some jet 4-vectors, sorts them by pT,
writes them to the tree, and then computes various possible
mjjs from different jet pairs. One of these is created directly
for the truth jets; for the partons they exist under PartonClusterer.

I still need to implement the actual "computes various possible
mjjs" part of this. What should be in place is the loading of
objects from the xAOD, the clustering, and the storing of particles
and truth/parton jets.