Implement the "best" mjj selection method in the Sherpa parton algorithm
The "best" method uses the truth jet mjj (from leading truth jets) and tries
to pick the partons which give a mjj that's closest to this. (It does the
same in the the other direction, too, which might be less interesting).

This caused me to do some re-architecturing. Clustering the partons and
processing the parton jets are now explicitly separate stages, to make
it easier to pass values around. I still think it's a bit awkward, but
it works, and is better than the way I'd done this before.