Add some parameters for the VBFInvSherpaTruth algorithm to runVBF.py
I added an argument group to our argparse parser, so these options
show up separately when looking at runVBF.py --help.

This provides knobs for changing the pT cuts that are applied when
selecting jets, the dR value used when running fastjet, and also
a switch to turn off fastjet altogether and just take the partons
as parton jets, for testing purposes.