Implement some minimal jet-parton matching in the new algorithm
The (un-cut) truth jet and parton jet collections are matched against
each other. In the output tree, I store the delta R of the closest
match, and the ID of the jet if the closest match was within 0.4

So for the parton jets, there's a branch of vectors that contain
the delta R of the closest truth jet, and the ID of that truth jet
*if* the delta R was less than 0.4.

I also store an integer that's the number of unmatched jets
in the event (where "unmatched" is defined as "there was no
corresponding truth/parton jet that had delta R <= 0.4).